requirements.txt
Flask>=2.2.0,<3.0.0
flask-cors>=4.0.0
gunicorn==23.0.0
gradio-client>=1.4.0
httpx>=0.27.0
Flask-Login>=0.6.3
Authlib>=1.3.0
Flask-SQLAlchemy>=3.1.0
Flask-Limiter>=3.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
praw>=7.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
cachetools>=5.0.0
user-agents>=2.2.0
google-generativeai>=0.8.0
tiktoken>=0.5.0
boto3>=1.34.0
Pillow>=10.0.0
Flask-WTF>=1.2.0
WTForms>=3.1.0